# packages/api/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clockwork Commerce API",
    "version": "1.0.0"
  },
  "components": {
...
    "securitySchemes": {
      "experienceSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "workos-session"
      }
    }
  },
  "paths": {
...
    "/v1/lifecycle/registrations": {
      "post": {
        "tags": [
          "lifecycle",
          "onboarding"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 255
            },
            "required": true,
            "name": "idempotency-key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "legalName": {
                    "type": "string",
                    "minLength": 2
                  },
                  "country": {
                    "type": "string",
                    "pattern": "^[A-Z]{2}$"
                  },
                  "registeredAddress": {
                    "type": "object",
                    "properties": {
                      "line1": {
                        "type": "string",
                        "minLength": 1
                      },
                      "line2": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string",
                        "minLength": 1
                      },
                      "region": {
                        "type": "string"
                      },
                      "postalCode": {
                        "type": "string",
                        "minLength": 1
                      },
                      "country": {
                        "type": "string",
                        "pattern": "^[A-Z]{2}$"
                      }
                    },
                    "required": [
                      "line1",
                      "city",
                      "postalCode",
                      "country"
                    ]
                  },
                  "relationshipRoles": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "direct_client",
                        "partner",
                        "end_client"
                      ]
                    },
                    "minItems": 1
                  },
                  "businessDomain": {
                    "type": "string",
                    "minLength": 3
                  },
                  "registrantEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "registrationToken": {
                    "type": "string",
                    "minLength": 32
                  },
                  "taxIds": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "jurisdiction": {
                          "type": "string",
                          "pattern": "^[A-Z]{2}$"
                        },
                        "value": {
                          "type": "string",
                          "minLength": 3
                        }
                      },
                      "required": [
                        "jurisdiction",
                        "value"
                      ]
                    },
                    "maxItems": 20
                  },
                  "billingContact": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string",
                        "format": "email"
                      }
                    },
                    "required": [
                      "name",
                      "email"
                    ]
                  },
                  "apContact": {
                    "type": "object",
                    "nullable": true,
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string",
                        "format": "email"
                      }
                    },
                    "required": [
                      "name",
                      "email"
                    ]
                  },
                  "invoiceDeliveryEmail": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": [
                  "legalName",
                  "country",
                  "registeredAddress",
                  "relationshipRoles",
                  "businessDomain",
                  "registrantEmail",
                  "registrationToken",
                  "taxIds",
                  "billingContact",
                  "apContact",
                  "invoiceDeliveryEmail"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lifecycle operation completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "minLength": 1
                    },
                    "eventType": {
                      "type": "string",
                      "minLength": 1
                    },
                    "signingUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  },
                  "required": [
                    "id",
                    "status"
                  ],
                  "additionalProperties": {
                    "nullable": true
                  }
                }
              }
            }
          },
          "403": {
            "description": "Account, role, MFA, or recent-authentication denial"
          },
          "503": {
            "description": "Lifecycle persistence adapter is not configured"
          }
        }
      }
    },
...
  }
}